Increased size of test graph.
Testing on a 40 core machine, the dynograph impl test would fail
when run with 34 cores or more. Edges were corrupted immediately
after they were added, leading to failed tests. My best guess is
that the vertex->rank mapping is broken when the number of vertices
is close to the number of processors. Increasing max_nv from 100 to 1000
fixed the issue.